+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
 - the `ImageData` class now uses the python-image-complete library to determine the file format rather than
   loading the image into memory in order to determine that
 - the `convert-image-format` filter now correctly creates a new container with the converted image data
+- the `to-coco-od` writer only allows sorting of categories when using predefined categories now
 
 
 0.0.1 (2024-05-06)